+++ include/CUDA_Bench/gemm/gemm_cutlass_launch_fp.cuh
@@ -70,7 +70,7 @@ int gemm_cutlass_launch_fp()
     gpuErrchk(cudaDeviceSynchronize());
     cudaEventRecord(time_stop, 0);
     cudaEventSynchronize(time_stop);
-    double elapsedTime;
+    float elapsedTime;
     cudaEventElapsedTime(&elapsedTime, time_start, time_stop);
     cudaEventDestroy(time_start);
     cudaEventDestroy(time_stop);
